requirements.txt
django
djangorestframework
django-heroku
gunicorn
psycopg2
